combine locale and language services into one
