+++ packages/blockchain-wallet-v4-frontend/src/components/DropdownLanguage/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 
-import * as languageService from 'services/LanguageService'
+import { languagesSortedByName } from 'services/LocalesService'
 import { SimpleDropdown } from 'blockchain-info-components/src/Dropdowns'
 import { actions, selectors } from 'data'
 
@@ -36,7 +36,7 @@ class DropdownLanguageContainer extends React.PureComponent {
 
 const mapStateToProps = state => ({
   currentLanguage: selectors.preferences.getLanguage(state),
-  languages: languageService.languagesSortedByName
+  languages: languagesSortedByName
 })
 
 const mapDispatchToProps = dispatch => ({
