+++ packages/blockchain-wallet-v4-frontend/src/components/Form/SelectBoxLanguages/index.js
@@ -4,7 +4,7 @@ import { assoc, curry, keys, map, reduce } from 'ramda'
 
 import { selectors } from 'data'
 import SelectBox from '../SelectBox'
-import * as languageService from 'services/LanguageService'
+import { languagesSortedByName } from 'services/LocalesService'
 
 const renameKeys = curry((keysMap, obj) =>
   reduce((acc, key) => assoc(keysMap[key] || key, obj[key], acc), {}, keys(obj))
@@ -23,9 +23,9 @@ class SelectBoxLanguages extends React.PureComponent {
   }
 }
 
-const mapStateToProps = (state, ownProps) => ({
+const mapStateToProps = state => ({
   language: selectors.preferences.getLanguage(state),
-  languages: languageService.languagesSortedByName
+  languages: languagesSortedByName
 })
 
 export default connect(mapStateToProps)(SelectBoxLanguages)
